Refactored the nmcli module
Got rid of every validation that can be done by the
remote host's NetworkManager.

Removed the 'is_exist...' methods in favor of
getter methods that simply return data - that way
if someone wants to know if a connection exists he can
simply check if it was returned.

Added a modify_device method so that devices properties
may also be altred.